Raise persistence and safety test quality
