+++ Tests/QuillCodePersistenceTests/PersistenceTestSupport.swift
@@ -1,0 +1,16 @@
+import XCTest
+
+class PersistenceTestCase: XCTestCase {
+    func makeTempDirectory() throws -> URL {
+        let url = URL(fileURLWithPath: NSTemporaryDirectory())
+            .appendingPathComponent("QuillCodePersistenceTests-\(UUID().uuidString)")
+        try FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
+        return url
+    }
+
+    func posixPermissions(at url: URL) throws -> Int {
+        let attributes = try FileManager.default.attributesOfItem(atPath: url.path)
+        let permissions = try XCTUnwrap(attributes[.posixPermissions] as? NSNumber)
+        return permissions.intValue & 0o777
+    }
+}
