+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -23,8 +23,8 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `test:QuillCodeCoreTests` | 3 | 524 | 98 | A+ | `CoreModelTests.swift` A+<br>`VerificationResultParserTests.swift` A+<br>`ModelTokenUsageTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 3 | 1108 | 97 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
 | `test:QuillCodeParityTests` | 145 | 13473 | 99 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A-<br>`ParityWorkspaceReviewSurfaceGateTests.swift` A-<br>`ParityMCPGateTests.swift` A- |
-| `test:QuillCodePersistenceTests` | 1 | 301 | 91 | A- | `PersistenceTests.swift` A- |
-| `test:QuillCodeSafetyTests` | 5 | 1246 | 95 | A | `SafetyGitPullRequestPolicyTests.swift` A-<br>`SafetyGeneralPolicyTests.swift` A-<br>`SafetyDownloadPolicyTests.swift` A- |
+| `test:QuillCodePersistenceTests` | 6 | 357 | 100 | A+ | `JSONProjectStoreTests.swift` A+<br>`ConfigStoreTests.swift` A+<br>`JSONAutomationStoreTests.swift` A+ |
+| `test:QuillCodeSafetyTests` | 5 | 1211 | 96 | A+ | `SafetyGitPullRequestPolicyTests.swift` A-<br>`SafetyDownloadPolicyTests.swift` A-<br>`SafetyGeneralPolicyTests.swift` A+ |
 | `test:QuillCodeToolsTests` | 26 | 3436 | 99 | A+ | `MCPStdioProberTests.swift` A<br>`ShellToolExecutorTests.swift` A<br>`FileToolExecutorTests.swift` A+ |
 | `test:QuillComputerUseKitTests` | 4 | 726 | 96 | A+ | `ComputerUseToolExecutorTests.swift` A-<br>`LinuxComputerUseBackendTests.swift` A<br>`LinuxComputerUseDetectionTests.swift` A+ |
 
@@ -1118,11 +1118,16 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 40 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceUIStateModelGateTests.swift` | no automated issues |
 | A+ | 100 | 25 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceViewCommandPlannerGateTests.swift` | no automated issues |
 | A+ | 100 | 83 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParityWorkspaceWorktreeGateTests.swift` | no automated issues |
-| A- | 91 | 301 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/PersistenceTests.swift` | 2 lines >120 chars; high duplicate-line ratio (28%) |
-| A- | 91 | 261 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyDownloadPolicyTests.swift` | 1 lines >120 chars; high duplicate-line ratio (29%) |
-| A- | 91 | 289 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyGeneralPolicyTests.swift` | 3 lines >120 chars; high duplicate-line ratio (53%) |
+| A+ | 100 | 93 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/ConfigStoreTests.swift` | no automated issues |
+| A+ | 100 | 38 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/FileSecretStoreTests.swift` | no automated issues |
+| A+ | 100 | 78 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/JSONAutomationStoreTests.swift` | no automated issues |
+| A+ | 99 | 116 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/JSONProjectStoreTests.swift` | 1 lines >120 chars |
+| A+ | 100 | 16 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/JSONThreadStoreTests.swift` | no automated issues |
+| A+ | 100 | 16 | `test:QuillCodePersistenceTests` | `Tests/QuillCodePersistenceTests/PersistenceTestSupport.swift` | no automated issues |
+| A- | 91 | 243 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyDownloadPolicyTests.swift` | 1 lines >120 chars; high duplicate-line ratio (31%) |
+| A+ | 99 | 209 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyGeneralPolicyTests.swift` | 1 lines >120 chars |
 | A- | 91 | 373 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyGitPullRequestPolicyTests.swift` | 2 lines >120 chars; high duplicate-line ratio (44%) |
-| A+ | 100 | 153 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyPolicyTestCase.swift` | no automated issues |
+| A+ | 100 | 216 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyPolicyTestCase.swift` | no automated issues |
 | A+ | 100 | 170 | `test:QuillCodeSafetyTests` | `Tests/QuillCodeSafetyTests/SafetyTests.swift` | no automated issues |
 | A+ | 99 | 191 | `test:QuillCodeToolsTests` | `Tests/QuillCodeToolsTests/FileToolExecutorTests.swift` | 7 lines >120 chars |
 | A+ | 99 | 38 | `test:QuillCodeToolsTests` | `Tests/QuillCodeToolsTests/GitBranchStatusIntegrationTests.swift` | 1 lines >120 chars |
